Fix: Set dynamic icon for weather notification
This commit updates the weather notification to display a dynamic icon based on the current weather conditions.

The `WeatherResponse` data class now includes an `icon` field within the `Weather` object. This field will store the icon code provided by the OpenWeatherMap API.

The `WeatherWorker` has been modified to extract this icon code from the API response and pass it to the `showWeatherNotification` function.

Finally, the `NotificationsHelper` class now includes a `getWeatherIconRes` function that maps the icon code to the corresponding drawable resource. This function is currently a placeholder and needs to be implemented with the actual icon resources. The `showWeatherNotification` function has been updated to use this new function to set the small icon for the notification.